Fixed the incompatible signal in Blockdiagram (closes ticket #39).
git-svn-id: svn://svn.code.sf.net/p/kactus2/code/trunk@40 f0aa6db5-7f46-43cc-a8a5-a31efdaafa67

+++ MCAPI/MCAPIFunctionSet.cpp
@@ -31,7 +31,7 @@ MCAPIFunctionSet::~MCAPIFunctionSet()
 //-----------------------------------------------------------------------------
 MCAPIFunctionDesc const& MCAPIFunctionSet::getAt(unsigned int index) const
 {
-    Q_ASSERT(index < m_functions.size());
+    Q_ASSERT(index < static_cast<unsigned int>(m_functions.size()));
     return m_functions.at(index);
 }
